Move conflict checking to its own file
